Move activity_finances out to its own endpoint
